add secret
